refactor: create an http layer folder
